Revert "Rename ReactServerFormatConfig to ReactFizzConfig (#26591)"
This reverts commit ffb8eaca5966fc7733bd0a23f4055e26d2cc59d7.
