+++ packages/react-dom/src/server/ReactDOMLegacyServerNodeStream.js
@@ -21,7 +21,7 @@ import {
 import {
   createResponseState,
   createRootFormatContext,
-} from 'react-dom-bindings/src/server/ReactFizzConfigDOMLegacy';
+} from 'react-dom-bindings/src/server/ReactDOMServerLegacyFormatConfig';
 
 import {Readable} from 'stream';
