+++ packages/react-server/src/ReactFizzNewContext.js
@@ -10,7 +10,7 @@
 import type {ReactContext} from 'shared/ReactTypes';
 
 import {REACT_SERVER_CONTEXT_DEFAULT_VALUE_NOT_LOADED} from 'shared/ReactSymbols';
-import {isPrimaryRenderer} from './ReactFizzConfig';
+import {isPrimaryRenderer} from './ReactServerFormatConfig';
 
 let rendererSigil;
 if (__DEV__) {
