+++ packages/react-server/src/ReactServerFormatConfig.js
