+++ packages/react-server/src/forks/ReactServerFormatConfig.custom.js
