+++ packages/react-server/src/forks/ReactServerFormatConfig.dom-bun.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from 'react-native-renderer/src/server/ReactFizzConfigNative';
+export * from 'react-dom-bindings/src/server/ReactDOMServerFormatConfig';
