+++ packages/react-server/src/forks/ReactServerFormatConfig.dom-legacy.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from 'react-dom-bindings/src/server/ReactFizzConfigDOM';
+export * from 'react-dom-bindings/src/server/ReactDOMServerLegacyFormatConfig';
