+++ packages/react-server/src/forks/ReactServerFormatConfig.dom-node.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from 'react-dom-bindings/src/server/ReactFizzConfigDOM';
+export * from 'react-dom-bindings/src/server/ReactDOMServerFormatConfig';
