+++ packages/react-server/src/forks/ReactServerFormatConfig.native-relay.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from 'react-dom-bindings/src/server/ReactFizzConfigDOM';
+export * from 'react-native-renderer/src/server/ReactNativeServerFormatConfig';
